pyproject.toml
[project]
name = "mlhard-1"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.0.4",
    "dvc>=3.62.0",
    "mlflow>=3.3.1",
    "pandas>=2.3.2",
    "scikit-learn>=1.7.1",
]
